[Decomp][Player] Cohesion split: extract vendor purchase system to PlayerVendor.cpp
Moves the contiguous vendor-purchase block out of Player.cpp into a
new file under src/game/Object/.

Player.cpp: 18,889 -> 18,423 lines (-466).
PlayerVendor.cpp: 549 lines, 4 methods.

Methods moved:
- TakeExtendedCost
- BuyItemFromVendorSlot, BuyCurrencyFromVendorSlot
- GetMaxPersonalArenaRatingRequirement

No header change; no signature change; no behaviour change.
